ui/requirements.txt
fastapi
uvicorn
tensorflow
scikit-learn
numpy
pillow
joblib
python-multipart
opencv-python
matplotlib
seaborn
pandas